chaps: fix linter errors
Fixed a number of errors reported by cpplint.

Added C++11 support to chaps.

Had to rework calls to thread-unsafe getpwnam, getpwuid and getgrnam
by replacing them with thread-safe .._r() versions of the functions.

BUG=None
TEST=FEATURES=test emerge-link platform2

Reviewed-on: https://chromium-review.googlesource.com/209114
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>

+++ chaps/session_impl.cc
@@ -831,7 +831,7 @@ bool SessionImpl::GenerateKeyPairSoftware(int modulus_bits,
                                           const string& public_exponent,
                                           Object* public_object,
                                           Object* private_object) {
-  if (public_exponent.length() > sizeof(unsigned long) ||
+  if (public_exponent.length() > sizeof(uint32_t) ||
       public_exponent.empty())
     return false;
   BIGNUM* e = ConvertToBIGNUM(public_exponent);
@@ -1316,4 +1316,4 @@ void SessionImpl::OperationContext::Clear() {
   parameter_.clear();
 }
 
-}  // namespace
+}  // namespace chaps
